requirements.txt
spotipy>=2.23.0
ytmusicapi>=1.3.0
python-dotenv>=1.0.0
